Move color impls to their own module, add lints Create `colors` module for Lab and Srgb, future colors Add lints to lib.rs Add lints to main.rs for binary Update dependencies
