+++ src/colors.rs
@@ -1,0 +1,4 @@
+mod kmeans;
+mod sort;
+
+pub use self::kmeans::MapColor;
